Add colorize_letters query flag to API.
If provided in the query parameters, this will use a hash of the URL passed in to color the letter icon used. The color is used when there is no icon of the requested size, and no "main color" could be found in smaller icons.

The idea is that this will prevent so many icons being grey when a large icon size is requested.